Shift function definitions from zmq_helper.h to zmq_helper.c
Compiling a library that uses ZMQ and linking that library to another program
runs into an issue due to the .h file having function definitions instead of
just declarations.  Move the definitions into their own .c file
